Remove deprecated configuration files and custom `rsync` action plugin

+++ inventory.py
@@ -7,6 +7,7 @@ import base64
 import yaml
 import copy
 from kubernetes import client, config
+from kubernetes.config import ConfigException
 
 """
 Teknoir Notebook custom dynamic inventory script for Ansible, in Python.
@@ -18,46 +19,24 @@ class TeknoirInventory(object):
     def __init__(self):
         self.inventory = {}
         self.read_cli_args()
-        self.inventory_path = self.args.output
 
         # Called with `--list`.
         if self.args.list:
             self.inventory = self.teknoir_inventory()
-            print(json.dumps(self.inventory, indent=4, sort_keys=True));
-        # Called with `--yaml` generate static inventory.
-        elif self.args.yaml:
-            self.inventory = self.teknoir_inventory()
-            inventory_temp = copy.deepcopy(self.inventory)
-            self.inventory_to_yaml = {'all': {'children': {}}}
-            self.inventory_to_yaml['all']['hosts'] = inventory_temp['_meta'].pop('hostvars')
-            inventory_temp.pop('_meta')
-            for group_name, group_dict in inventory_temp.items():
-                self.inventory_to_yaml['all']['children'][group_name] = {
-                    'hosts': {host: {} for host in group_dict['hosts']},
-                    'vars': group_dict['vars']
-                }
-
-            inventory_file = f'{self.inventory_path}/inventory.yaml'
-            with open(inventory_file, 'w') as outfile:
-                outfile.write(yaml.dump(self.inventory_to_yaml, default_flow_style=False))
         # Called with `--host [hostname]`.
         elif self.args.host:
             # Not implemented, since we return _meta info `--list`.
             self.inventory = self.teknoir_inventory()
-            print(json.dumps(self.inventory, indent=4, sort_keys=True));
         # If no groups or vars are present, return empty inventory.
         else:
             self.inventory = self.empty_inventory()
-            print(json.dumps(self.inventory, indent=4, sort_keys=True));
 
-
+        print(json.dumps(self.inventory, indent=4, sort_keys=True))
 
     def decode(self, s):
         return base64.b64decode(s.encode('utf-8')).decode('utf-8')
 
     def teknoir_inventory(self):
-        namespace = os.getenv('NAMESPACE', "namespace does not exist")
-
         try:
             config.load_kube_config()
         except config.ConfigException:
@@ -66,21 +45,35 @@ class TeknoirInventory(object):
             except config.ConfigException:
                 raise Exception("Could not configure kubernetes python client")
 
+        contexts, current_context = config.list_kube_config_contexts()
+        if not contexts or len(contexts) < 2:
+            raise Exception("No valid kube config contexts found")
+
+        def get_domain(value):
+            return {
+                'gke_teknoir_us-central1-c_teknoir-cluster': 'teknoir.cloud',
+                'gke_teknoir-poc_us-central1-c_teknoir-dev-cluster': 'teknoir.dev',
+            }.get(value, 'teknoir.cloud')
+        domain = get_domain(current_context['context']['cluster'])
+        namespace = current_context['context'].get('namespace', os.environ.get('NAMESPACE', 'default'))
+
         custom_api = client.CustomObjectsApi()
-        devices = custom_api.list_namespaced_custom_object(group="kubeflow.org",
+        devices = custom_api.list_namespaced_custom_object(group="teknoir.org",
                                                            version="v1",
-                                                           plural="devices",
-                                                           namespace=namespace)
+                                                           namespace=namespace,
+                                                           plural="devices")
 
         inventory = {
             '_meta': {
                 'hostvars': {}
             }
         }
-        ssh_port = 2200
+
         for device in devices['items']:
             ansible_group = device["metadata"]["namespace"].replace('-', '_').replace('.', '_')
-            path = f'{self.inventory_path}/{ansible_group}/'
+
+            home_dir = os.path.expanduser("~")
+            inventory_path = os.path.join(home_dir, ".ansible", "tmp", "inv", ansible_group)
             hostname = f'{device["metadata"]["name"]}'
 
             if ansible_group not in inventory:
@@ -88,7 +81,7 @@ class TeknoirInventory(object):
                     'hosts': [],
                     'vars': {}
                 }
-                os.makedirs(path, exist_ok=True)
+                os.makedirs(inventory_path, exist_ok=True)
             inventory[ansible_group]['hosts'].append(hostname)
 
             for label, value in device["metadata"]["labels"].items():
@@ -102,27 +95,51 @@ class TeknoirInventory(object):
                     }
                 inventory[additional_group]['hosts'].append(hostname)
 
-            private_key_file = f'{path}{device["metadata"]["name"]}.pem'
-            with open(private_key_file, 'w') as outfile:
-                outfile.write(self.decode(device['spec']['keys']['data']['rsa_private']))
-            os.chmod(private_key_file, 0o600)
+            private_key_file = os.path.join(inventory_path, f'{device["metadata"]["name"]}.pem')
+            if not os.path.isfile(private_key_file):
+                with open(private_key_file, 'w') as outfile:
+                    outfile.write(self.decode(device['spec']['keys']['data']['rsa_private']))
+                os.chmod(private_key_file, 0o400)
 
+            # print(json.dumps(device, indent=4, sort_keys=True))
+
+            tunnel_port = str(random.randint(1024, 64511))
+            tunnel_opened = False
+            if ('remote_access' in device['subresources']['status'] and
+                'active' in device['subresources']['status']['remote_access'] and
+                'port' in device['subresources']['status']['remote_access']):
+                tunnel_opened = device['subresources']['status']['remote_access']['active']
+                tunnel_port = device['subresources']['status']['remote_access']['port']
+
+            if ('data' not in device['spec']['keys'] or
+                'username' not in device['spec']['keys']['data'] or
+                'userpassword' not in device['spec']['keys']['data']):
+                continue
+
+            deadendhost = f'deadend.{namespace}'
+            deadendport = 22
+            username = self.decode(device['spec']['keys']['data']['username'])
+            userpassword = self.decode(device['spec']['keys']['data']['userpassword'])
+            pcmd = f"ssh -o UserKnownHostsFile=/dev/null -o StrictHostKeyChecking=no -o ExitOnForwardFailure=yes -o ServerAliveInterval=60 -i {private_key_file} -N -W %h:%p teknoir@{deadendhost} -p {deadendport}"
             inventory['_meta']['hostvars'][hostname] = {
-                'ansible_connection': 'ssh',
-                'ansible_port': self.decode(device['spec']['keys']['data']['tunnel']),
-                'ansible_host': 'localhost',
-                'ansible_user': self.decode(device['spec']['keys']['data']['username']),
-                'ansible_sudo_pass': self.decode(device['spec']['keys']['data']['userpassword']),
-                'ansible_become': 'yes',
+                'ansible_namespace': device["metadata"]["namespace"],
+                'ansible_port': tunnel_port,
+                'ansible_host': '127.0.0.1',
+                'ansible_user': username,
+                'ansible_sudo_pass': userpassword,
                 'ansible_become_user': 'root',
-                'ansible_become_pass': self.decode(device['spec']['keys']['data']['userpassword']),
+                'ansible_become_pass': userpassword,
+                'ansible_become_flags': '-E',
                 'ansible_ssh_private_key_file': private_key_file,
+                'ansible_ssh_args': f'-o ForwardAgent=yes -o ProxyCommand="{pcmd}"',
                 'ansible_python_interpreter': '/usr/bin/python3',
                 'ansible_ssh_retries': 20,
                 'ansible_kubectl_namespace': device["metadata"]["namespace"],
-                'ansible_teknoir_device': device['metadata']['name']
+                'ansible_teknoir_tunnel_port': tunnel_port,
+                'ansible_teknoir_tunnel_open': tunnel_opened,
+                'ansible_teknoir_device': device['metadata']['name'],
+                'ansible_teknoir_domain': domain,
             }
-            ssh_port = ssh_port + 1
         return inventory
 
     # Empty inventory for testing.
@@ -131,12 +148,9 @@ class TeknoirInventory(object):
 
     # Read the command line args passed to the script.
     def read_cli_args(self):
-        inventory_path = f'{os.getcwd()}/.inventory'
         parser = argparse.ArgumentParser()
         parser.add_argument('--list', action='store_true')
-        parser.add_argument('--yaml', action='store_true')
         parser.add_argument('--host', action='store')
-        parser.add_argument('--output', default=inventory_path)
         self.args = parser.parse_args()
